refactor(config): refactor processing module to use pydantic config
